jdk8

+++ genos/src/test/java/genos/ExampleUnitTest.kt
@@ -22,11 +22,11 @@ import org.junit.Test
 /**
  * Example local unit test, which will execute on the development machine (host).
  *
- * @see [Testing documentation](https://d.android.com/tools/testing)
+ * See [testing documentation](http://d.android.com/tools/testing).
  */
 class ExampleUnitTest {
     @Test
     fun addition_isCorrect() {
-        assertEquals(4, (2 + 2).toLong())
+        assertEquals(4, 2 + 2)
     }
 }
